refactor: make services class based
